+++ app/src/main/java/com/pklproject/checkincheckout/api/models/apimodel/Persentase.kt
@@ -5,7 +5,7 @@ import com.google.gson.annotations.SerializedName
 
 data class Persentase(
     @SerializedName("persentase_kehadiran")
-    val persentaseKehadiran: Int?,
+    val persentaseKehadiran: Float?,
     @SerializedName("persentase_ketidakhadiran")
-    val persentaseKetidakhadiran: Int?
+    val persentaseKetidakhadiran: Float?
 )
